requirements.txt
absl-py==2.3.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.4
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.1.0
narwhals==1.43.1
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==5.29.5
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.46.0
streamlit-autorefresh==1.0.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.63
